agent/requirements.txt
# Core LLM + LangGraph
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-mistralai>=0.1.0
openai>=1.30.0

# Web server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
mangum>=0.17.0

# Git + GitHub API
gitpython>=3.1.43
PyGithub>=2.3.0

# Data validation
pydantic>=2.7.0

# Database
pymongo>=4.7.0

# Environment
python-dotenv>=1.0.0

# HTTP client (used by openai / langchain)
httpx>=0.27.0
